+++ scripts/snippets/enable-ssl.sh
@@ -1,4 +1,4 @@
-# variables: $domain, $email
+# variables: $domain, $email, $port
 
 if egrep 'ssl\s*on;' /etc/nginx/sites-available/$domain >/dev/null ; then
 	echo "It looks like SSL is already setup for $domain"
@@ -25,5 +25,8 @@ else
 fi
 
 sudo cp /srv/.templates/$template /etc/nginx/sites-available/${domain}
-sudo sed -i -e s/{{domain}}/${domain}/g /etc/nginx/sites-available/${domain}
+sudo sed -i\
+	-e s/{{domain}}/${domain}/g\
+	-e s/{{port}}/${port}/g\
+	/etc/nginx/sites-available/${domain}
 sudo systemctl restart nginx