+++ templates/post-receive-static.sh
@@ -16,6 +16,19 @@ cleanup() {
 trap cleanup EXIT
 
 log '---- post-receive script started! ----'
+
+# only deploy to the master branch
+while read old new ref; do
+    branch=$(git rev-parse --symbolic --abbrev-ref $ref)
+    if [[ $branch != "master" ]]; then
+        log "'$branch' is not 'master'. A build is only triggered when pushing the master branch."
+		log "'$branch' was successfully pushed, but project was not built."
+		log
+		log 'Have a great day!'
+        exit 0
+    fi
+done
+
 log "cloning project to '$TMP_REPO'..."
 
 git clone $(pwd) $TMP_REPO