+++ templates/post-receive.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 SITE_DIR=/srv/{{site}}
-WAR_TARGET_LOCATION=/opt/tomcat/{{site}}/ROOT.war
+JAR_TARGET_LOCATION=/srv/{{site}}/app.jar
 
 TMP_REPO=$(mktemp -d)
 
@@ -17,6 +17,19 @@ cleanup() {
 trap cleanup EXIT
 
 log '---- post-receive script started! ----'
+
+# only deploy to the master branch
+while read old new ref; do
+    branch=$(git rev-parse --symbolic --abbrev-ref $ref)
+    if [[ $branch != "master" ]]; then
+        log "'$branch' is not 'master'. A build is only triggered when pushing the master branch."
+		log "'$branch' was successfully pushed, but project was not built."
+		log
+		log 'Have a great day!'
+        exit 0
+    fi
+done
+
 log "cloning project to '$TMP_REPO'..."
 
 git clone $(pwd) $TMP_REPO
@@ -29,24 +42,24 @@ if [[ -f $SITE_DIR/config ]]; then
 		log "Copying $source file to $destination..."
 		cp $SITE_DIR/$source $TMP_REPO/$destination
 	else
-		log "Configuration file found '${SITE_DIR}/config', but $source and $destination are not set."
+		log "Configuration file found '${SITE_DIR}/config', but \$source and \$destination are not set."
 		log 'Nothing copied. Continuing...'
 	fi
 else
 	log "No configuration file ($SITE_DIR/config) found. Continuing..."
 fi
 
-if [[ -f .build_config ]]; then
-	log 'Found ".build_config" file! Building based on this file...'
-	source .build_config
+if [[ -f .cods ]]; then
+	log 'Found ".cods" file! Building based on this file...'
+	source .cods
 
 	if [[ -z $BUILD_COMMAND ]]; then
-		log '$BUILD_COMMAND not set! (Check the .build_config file)'
+		log '$BUILD_COMMAND not set! (Check the .cods file)'
 		log 'Aborting...'
 		exit 1
 	fi
-	if [[ -z $WAR_FILE ]]; then
-		log '$WAR_FILE not set! (Check the .build_config file)'
+	if [[ -z $JAR_FILE ]]; then
+		log '$JAR_FILE not set! (Check the .cods file)'
 		log 'Aborting...'
 		exit 1
 	fi
@@ -65,26 +78,28 @@ if [[ -f .build_config ]]; then
 		log 'Aborting...'
 		exit 1
 	fi
-	if [[ ! -f $WAR_FILE ]]; then
-		log "Build was successful, but war file: '$WAR_FILE' was not found!"
+	if [[ ! -f $JAR_FILE ]]; then
+		log "Build was successful, but jar file: '$JAR_FILE' was not found!"
 		log 'Aborting...'
 		exit 1
 	fi
 
-	log "Build success! Deploying $WAR_FILE to $WAR_TARGET_LOCATION..."
-	rm -f $WAR_TARGET_LOCATION
-	mv $WAR_FILE $WAR_TARGET_LOCATION
+	log "Build success! Deploying $JAR_FILE to $JAR_TARGET_LOCATION..."
+	rm -f $JAR_TARGET_LOCATION
+	mv $JAR_FILE $JAR_TARGET_LOCATION
+
+	sudo service {{site}} restart
 
 	log '{{site}} deployed!'
 
 elif [[ -f install.sh ]]; then
 	log 'Found "install.sh"! Running...'
 	export SITE_DIR
-	export WAR_TARGET_LOCATION
+	export JAR_TARGET_LOCATION
 	export TMP_REPO
 	bash install.sh
 else
-	log 'No ".build_config" file or "install.sh" file found.'
+	log 'No ".cods" file or "install.sh" file found.'
 fi
 
 log '--------------------------------------------------'