+++ tests/sample-sites/static-with-build/install.sh
@@ -1,0 +1,4 @@
+
+set -u
+
+echo 'Static Site with a build step is Working!' > $PUBLIC_DIR/index.html